Move pullAndUpdate to course.ts (avoid circular dependency)
